Conforming to POSIX (IEEE 1003.1-2001)

+++ include/sys/types.h
@@ -9,8 +9,8 @@
 typedef int ssize_t;
 typedef unsigned int size_t;
 
-typedef short uid_t; /* user id */
-typedef short gid_t; /* group id */
+typedef int uid_t; /* user id */
+typedef int gid_t; /* group id */
 typedef int pid_t; /* process id */
 typedef unsigned char flag_t; /* 0/1 logical flag */
